Add unit tests for the new YoutubeAtom
